# Make it easier to see test performance