--- a/testing/file_test/run_test.cpp
+++ b/testing/file_test/run_test.cpp
@@ -164,11 +164,13 @@ auto RunTestFile(const FileTestBase& test_base, bool dump_output,
   llvm::raw_svector_ostream output_stream(test_file.actual_stdout);
   llvm::raw_svector_ostream error_stream(test_file.actual_stderr);
 
+  Timer timer;
   ErrorOr<FileTestBase::RunResult> run_result =
       dump_output ? test_base.Run(test_args_ref, fs, input_stream, llvm::outs(),
                                   llvm::errs())
                   : test_base.Run(test_args_ref, fs, input_stream,
                                   output_stream, error_stream);
+  test_file.run_elapsed_ms = timer.elapsed_ms();
 
   // Ensure stdout/stderr are always fetched, even when discarded on error.
   if (test_file.capture_console_output) {