--- a/toolchain/autoupdate_testdata.py
+++ b/toolchain/autoupdate_testdata.py
@@ -42,6 +42,8 @@ def main() -> None:
     # Parse arguments.
     parser = argparse.ArgumentParser(__doc__)
     parser.add_argument("--non-fatal-checks", action="store_true")
+    parser.add_argument("--print_slowest_tests", help="Forwarded to file_test")
+    parser.add_argument("--threads", help="Forwarded to file_test")
     parser.add_argument("files", nargs="*")
     args = parser.parse_args()
 
@@ -65,6 +67,10 @@ def main() -> None:
         "--",
         "--autoupdate",
     ]
+    if args.print_slowest_tests:
+        argv += ["--print_slowest_tests", args.print_slowest_tests]
+    if args.threads:
+        argv += ["--threads", args.threads]
     # Support specifying tests to update, such as:
     # ./autoupdate_testdata.py lex/**/*
     if args.files: