--- a/toolchain/autoupdate_testdata.py
+++ b/toolchain/autoupdate_testdata.py
@@ -42,7 +42,9 @@ def main() -> None:
     # Parse arguments.
     parser = argparse.ArgumentParser(__doc__)
     parser.add_argument("--non-fatal-checks", action="store_true")
-    parser.add_argument("--print_slowest_tests", help="Forwarded to file_test")
+    parser.add_argument(
+        "--print_slowest_tests", default=0, help="Forwarded to file_test"
+    )
     parser.add_argument("--threads", help="Forwarded to file_test")
     parser.add_argument("files", nargs="*")
     args = parser.parse_args()
@@ -66,9 +68,9 @@ def main() -> None:
         "//toolchain/testing:file_test",
         "--",
         "--autoupdate",
+        "--print_slowest_tests",
+        str(args.print_slowest_tests),
     ]
-    if args.print_slowest_tests:
-        argv += ["--print_slowest_tests", args.print_slowest_tests]
     if args.threads:
         argv += ["--threads", args.threads]
     # Support specifying tests to update, such as: